move UI init code
